Move ErrorToast component to ErrorContainer component
