+++ admin/resources/js/components/ErrorContainer.tsx
@@ -1,4 +1,5 @@
-import React, { useReducer } from "react";
+import React, { useCallback, useContext, useReducer } from "react";
+import Alert from "./H2Components/Alert";
 
 type ErrorState = {
   errorQueue: string[];
@@ -27,14 +28,41 @@ interface ErrorContextProps {
 
 export const ErrorContext = React.createContext<ErrorContextProps>({
   state: initialState,
-  // eslint-disable-next-line @typescript-eslint/no-empty-function
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  dispatch: (action) => {},
+  dispatch: () => {
+    /* returns nothing */
+  },
 });
 
-export const ErrorContainer: React.FunctionComponent = ({
-  children,
-}): React.ReactElement => {
+export const ErrorToast: React.FC = () => {
+  const { state, dispatch } = useContext(ErrorContext);
+  const dismiss = useCallback(() => dispatch({ type: "pop" }), [dispatch]);
+
+  // This toast will render the first error in the queue, if any.
+  const currentError = state.errorQueue.length > 0 ? state.errorQueue[0] : null;
+
+  return (
+    <>
+      {currentError !== null && (
+        <Alert
+          color="pink"
+          position="toast"
+          dismissBtn={
+            <Alert.DismissBtn onClick={dismiss} aria-label="Dismiss">
+              Dismiss
+            </Alert.DismissBtn>
+          }
+        >
+          <p>
+            <strong>Something went wrong!</strong>
+          </p>
+          <p>{currentError}</p>
+        </Alert>
+      )}
+    </>
+  );
+};
+
+export const ErrorContainer: React.FC = ({ children }): React.ReactElement => {
   const [state, dispatch] = useReducer(errorReducer, initialState);
 
   return (
